DELIA-43983 : [Thunder]-System Plugin Issues
Reason for change: Updated getCachedValue, setCachedValue, cacheContains, removeCacheKey, getLastDeepSleepReason, enableXREConnectionRetention.
Test Procedure: Refer ticket
Risks: Low

+++ SystemServices/SystemServices.cpp
@@ -1643,25 +1643,29 @@ namespace WPEFramework {
 
         /***
          * @brief : To get cashed value .
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>"}}}
+         * @param1[in]  : {"params":{"key":"<string>"}}
          * @param2[out] : {"result":{"<cachekey>":"<string>","success":<bool>}}
          * @return      : Core::<StatusCode>
          */
         uint32_t SystemServices::getCachedValue(const JsonObject& parameters,
-                JsonObject& response)
+			JsonObject& response)
         {
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
+            bool retStat = false;
+            std::string key = parameters["key"].String();
             LOGWARN("key: %s\n", key.c_str());
-            response[(key.c_str())] = m_cacheService.getValue(key).String();
-            returnResponse(true);
+            if (key.length()) {
+                response[(key.c_str())] = m_cacheService.getValue(key).String();
+                retStat = true;
+            } else {
+                populateResponseWithError(SysSrv_UnSupportedFormat, response);
+            }
+            returnResponse(retStat);
         }
 
         /***
          * @brief : To set cache value.
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>",
-         *                 "cacheValue":<double>}}}
+         * @param1[in]  : {"params":{"key":"<string>",
+         *                 "cacheValue":<double>}}
          * @param2[out] : {"jsonrpc":"2.0","id":3,"result":{"success":<bool>}}
          * @return      : Core::<StatusCode>
          */
@@ -1669,22 +1673,24 @@ namespace WPEFramework {
                 JsonObject& response)
         {
             bool retStat = false;
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
-            std::string value = param["cacheValue"].String();
+            std::string key = parameters["key"].String();
+            std::string value = parameters["value"].String();
 
-            if (m_cacheService.setValue(key, value)) {
-                retStat = true;
+            if (key.length() && value.length()) {
+                if (m_cacheService.setValue(key, value)) {
+                    retStat = true;
+                } else {
+                    LOGERR("Accessing m_cacheService.setValue failed\n.");
+                }
             } else {
-                LOGERR("Accessing m_cacheService.setValue failed\n.");
+                populateResponseWithError(SysSrv_UnSupportedFormat, response);
             }
             returnResponse(retStat);
         }
 
         /***
          * @brief : To check if key value present in cache.
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>"}}}
+         * @param1[in]  : {"params":{"key":"<string>"}}
          * @param2[out] : {"jsonrpc":"2.0","id":3,"result":{"success":<bool>}}
          * @return      : Core::<StatusCode>
          */
@@ -1692,27 +1698,22 @@ namespace WPEFramework {
                 JsonObject& response)
         {
             bool retStat = false;
-            if (parameters.HasLabel("key")) {
-                std::string key = parameters["key"].String();
-                if (key.length()) {
-                    if (m_cacheService.contains(key)) {
-                        retStat = true;
-                    } else {
-                        LOGERR("Accessing m_cacheService.contains; no matching key '%s'\n.", key.c_str());
-                        populateResponseWithError(SysSrv_KeyNotFound, response);
-                    }
+            std::string key = parameters["key"].String();
+            if (key.length()) {
+                if (m_cacheService.contains(key)) {
+                    retStat = true;
                 } else {
-                    populateResponseWithError(SysSrv_UnSupportedFormat, response);
+                    LOGERR("Accessing m_cacheService.contains failed\n.");
                 }
             } else {
-                populateResponseWithError(SysSrv_MissingKeyValues, response);
+                populateResponseWithError(SysSrv_UnSupportedFormat, response);
             }
             returnResponse(retStat);
         }
 
         /***
          * @brief : To delete the key value present in cache.
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>"}}}
+         * @param1[in]  : {"params":{"key":"<string>"}}
          * @param2[out] : {"jsonrpc":"2.0","id":3,"result":{"success":<bool>}}
          * @return      : Core::<StatusCode>
          */
@@ -1720,13 +1721,16 @@ namespace WPEFramework {
                 JsonObject& response)
         {
             bool retStat = false;
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
-            if (m_cacheService.remove(key)) {
-                retStat = true;
+            std::string key = parameters["key"].String();
+            if (key.length()) {
+                if (m_cacheService.remove(key)) {
+                    retStat = true;
+                } else {
+                    LOGERR("Accessing m_cacheService.remove failed\n.");
+                    populateResponseWithError(SysSrv_Unexpected, response);
+                }
             } else {
-                LOGERR("Accessing m_cacheService.remove failed\n.");
+                populateResponseWithError(SysSrv_UnSupportedFormat, response);
             }
             returnResponse(retStat);
         }
@@ -1802,20 +1806,22 @@ namespace WPEFramework {
             string reason;
 
             if (Utils::fileExists(STANDBY_REASON_FILE)) {
-                    std::ifstream inFile(STANDBY_REASON_FILE);
-                    if (inFile) {
-                        std::getline(inFile, reason);
-                        inFile.close();
-                        retAPIStatus = true;
-                    } else {
-                        populateResponseWithError(SysSrv_FileAccessFailed, response);
-                    }
+                std::ifstream inFile(STANDBY_REASON_FILE);
+                if (inFile) {
+                    std::getline(inFile, reason);
+                    inFile.close();
+                    retAPIStatus = true;
                 } else {
-                    populateResponseWithError(SysSrv_FileNotPresent, response);
+                    populateResponseWithError(SysSrv_FileAccessFailed, response);
                 }
+            } else {
+                populateResponseWithError(SysSrv_FileNotPresent, response);
+            }
 
-                if (retAPIStatus && reason.length()) {
-                    response["lastDeepSleepReason"] = reason;
+            if (retAPIStatus && reason.length()) {
+                response["reason"] = reason;
+            } else {
+                response["reason"] = "";
             }
             returnResponse(retAPIStatus);
         }
@@ -2086,7 +2092,7 @@ namespace WPEFramework {
 
         /***
          * @brief : Enables XRE Connection Retension option.
-         * @param1[in]  : {"params":{"param":<bool>}}
+         * @param1[in]  : {"params":{"enable":<bool>}}
          * @param2[out] : "result":{"success":<bool>}
          * @return      : Core::<StatusCode>
          */
@@ -2096,11 +2102,11 @@ namespace WPEFramework {
             bool enable = false, retstatus = false;
             int status = SysSrv_Unexpected;
 
-                enable = parameters["param"].Boolean();
-                if ((status = enableXREConnectionRetentionHelper(enable)) == SysSrv_OK) {
-                    retstatus = true;
-                } else {
-                    populateResponseWithError(status, response);
+            enable = parameters["enable"].Boolean();
+            if ((status = enableXREConnectionRetentionHelper(enable)) == SysSrv_OK) {
+                retstatus = true;
+            } else {
+                populateResponseWithError(status, response);
             }
             returnResponse(retstatus);
         }
